requirements.txt
amqp==5.1.1
APScheduler==3.9.1.post1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.50.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
filelock==3.8.0
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
packaging==21.3
pipenv==2022.11.23
pipreqs==0.4.11
platformdirs==2.5.4
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
schedule==1.1.0
sendgrid==6.9.7
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tomli==2.0.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
vine==5.0.0
virtualenv==20.16.7
virtualenv-clone==0.5.7
waitress==2.1.2
wcwidth==0.2.5
yarg==0.1.9
